modify(post): 게시물 검색을 v2 목록 API에 통합
별도 검색 엔드포인트를 제거하고 open-api/v2/posts의 선택적 keyword 필터로 검색 계약을 통합한다. 검색 동작과 마이그레이션 검증 테스트를 각 기존 책임의 테스트로 옮긴다.
