+++ src/main/kotlin/com/techtaurant/mainserver/post/application/PostListReadService.kt
@@ -53,7 +53,6 @@ class PostListReadService(
      * @param authorId 작성자 필터 (null이면 전체 조회)
      * @param categoryId 카테고리 필터 (null이면 전체, authorId 지정 시에만 적용)
      * @param tagIds 태그 UUID 필터 (여러 개 전달 시 OR 조건)
-     * @param keyword 제목 또는 본문 부분 일치 검색어 (null이면 미적용, 대소문자 무시)
      * @return 커서 기반 페이지 응답
      */
     fun getPosts(
@@ -65,7 +64,6 @@ class PostListReadService(
         authorId: UUID? = null,
         categoryId: UUID? = null,
         tagIds: List<UUID>? = null,
-        keyword: String? = null,
     ): CursorPageResponse<PostListItemResponse> {
         val postPage =
             getPostPage(
@@ -77,7 +75,6 @@ class PostListReadService(
                 authorId = authorId,
                 categoryId = categoryId,
                 tagIds = tagIds,
-                keyword = keyword,
             )
         val content = postPage.content
 
@@ -124,6 +121,7 @@ class PostListReadService(
         authorId: UUID? = null,
         categoryId: UUID? = null,
         tagIds: List<UUID>? = null,
+        keyword: String? = null,
     ): CursorPageResponse<PostContentListItemResponse> {
         val postPage =
             getPostPage(
@@ -135,6 +133,7 @@ class PostListReadService(
                 authorId = authorId,
                 categoryId = categoryId,
                 tagIds = tagIds,
+                keyword = keyword,
             )
 
         return CursorPageResponse(
